requirements.txt
flask
python-chess
